Completed Pathfinding
The player now controls the big fish by left clicking on the screen. There are still a couple of small bugs with the pathfinding but nothing with the code. Just things like it uses too narrow a raycast so sometimes it trys to cut between tiny gaps, and when it is right on top of a wall sometimes it can get stuck for a moment.